Refactor and optimize parse-string
Replace (shell "$'\xHH'") with (strings-from-bytes N).  This is much faster,
and it makes the definitions of special characters circular.  String parsing
relies in string.scm for the definition of `\xHH`, which in turn defines
them using `\xHH` sequences (supplied by the golden compiler).

Add test for `\x00`.

Fix handling of `\x` after `;` inside a string.