Remove support for `use` directive
Previously, `(use MODULE)` would be used to load a module that exports
executable macros (defined with `defmacro`).  This would make only the
`defmacro` definitions available during compile-time, but MODULE's other
exports would not be available at run-time and MODULE would not be
included in the generated executable.

Now this is done with `(require MODULE)`.  This loads compile-time
(`defmacro`) and run-time (all other) exports from MODULE.  When
`require` is processed by the compiler and the required module is found
to export executable macros, that module will be loaded into the
compiler.

Executable macros still cannot be used in the module in which they are
defined.  They can only be used by a module that requires the module
that defines them.