Save memo DB on memo-drop
After errors are reported, the compiler may continue for quite some time
before exiting the memo session, and if the compiler is killed by the
user, the memo DB does not get saved ... unless we save it explicitly.
memo-drop is a good time to save the DB.  It is an indication of an
error and that terminaton is likely, and it cannot slow down the
"happy" path (when compliation succeeds).