Blobify writes
memo-write-file saves data first as a BLOB and then copies the
BLOB over the specified file name.  The "copy BLOB over file"
step is wrapped with memo-io.  On playback, when the destination
file hash does not match, it simply repeats the copy.  (We assume
BLOBs will persist as long as the memo DB.)