Generalized destructuring
Where symbols can be bound, more complex destructuring forms may also be
used in all of the contexts in which variables are bound, except for `case`
forms.

The following destructuring forms ("targets") are supported:

 TARGET :=  SYMBOL
         |  [TARGET...]
         |  [TARGET... REST]
         |  { =TARGET: TARGET, ... }
         |  { =TARGET: TARGET, ... , `...`:SYMBOL }
         |  { KEY: TARGET, ...}

REST is a symbol consisting of `...` followed by a variable name.

KEY is a symbol or a string literal.  The associated TARGET is bound to the
value associated stored under that KEY.

Consolidate target processing in gen.scm

Rename "pattern" to "target".