Fix top-level assignment with literal value ending in "\"
Assignments in "file syntax" (which includes all top-level define's and
assignments) compile to "VAR = VALUE", "VAR := VALUE", or "define VAR
... endef".  We did not account for the following quirky behavior that
occurs when ":=" or "=" lines end in a backslash:

  \      -->  <empty>
  \\     -->  \\
  \\\    -->  \<space>
  \\\\   -->  \\\\
  \\\\\  -->  \\<space>
  \\\\\\ -->  \\\\\\

Note there is no way to represent a value that ends in a single backslash by
simply enoding it with another backslash.

To avoid this we append a single-character `#` comment when values end in
`\`, since the handling of backslashes immediately preceding a "#" appears
to work as documented.